Update Hibari scripts to support multi-node config.

+++ bin/hibari-reset-ossforum.sh
@@ -1,8 +1,7 @@
 #!/bin/sh
 # -*- mode:shell-script; coding:utf-8 -*-
 
-# ALL_NODES=(hibari1 hibari2 hibari3)
-ALL_NODES=(hibari)
+ALL_NODES=(hibari1 hibari2 hibari3)
 
 die() {
     echo $@
@@ -19,7 +18,11 @@ delete_node_data() {
            cd $HIBARI_HOME/rel/$NODE; \
                rm -rf data/* || \
                die "node $NODE reset failed"
-           rm Schema.local
+
+           if [ $I -eq 0 ]; then
+               rm Schema.local
+           fi
+
            echo deleted data on $NODE
        ) &
    done
